first couple of tests are now working
